requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
starlette==0.27.0

# Validation & Settings
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.10.1
email-validator==2.3.0

# Database
pymongo==4.6.0
dnspython==2.6.1

# Authentication
PyJWT==2.8.0

# Payment Processing
stripe==11.2.0

# File Upload
python-multipart==0.0.6
pillow==12.0.0

# HTTP & Networking
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Environment & Configuration
python-dotenv==1.1.1

# Utilities
click==8.2.1
colorama==0.4.6
h11==0.16.0
sniffio==1.3.1
anyio==3.7.1
exceptiongroup==1.3.0

# Type Checking
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0

# Packaging
packaging==25.0

# PayPal (Official SDK for Orders API v2 - compatible with @paypal/react-paypal-js)
paypal-server-sdk==2.2.0